clean up code, renaming and refactoring code
