+++ lib/helpers.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+import 'icon_text_data.model.dart';
+
+final List<IconTextData> goalOfApp = [
+  IconTextData(const Icon(Icons.pets),
+      " recently got a dog and are looking for friends for walks;\n"),
+  IconTextData(const Icon(Icons.pets),
+      " have moved to another city and do not know anyone; \n"),
+  IconTextData(const Icon(Icons.pets),
+      " have certain characteristics in their dog that cannot walk with any dog; \n"),
+  IconTextData(const Icon(Icons.pets),
+      " are looking for dogs for mating of certain breeds; \n"),
+  IconTextData(
+      const Icon(Icons.pets), " seek socialization of dogs and people. \n"),
+];
