+++ lib/icon_text_data.model.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class IconTextData {
+  final Icon icon;
+  final String text;
+
+  IconTextData(this.icon, this.text);
+}
