+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:dog_friends/welcome_screen.dart';
+import 'package:dog_friends/screens/welcome_screen.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -32,7 +32,7 @@ class MyApp extends StatelessWidget {
         colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
         useMaterial3: true,
       ),
-      home: const StartScreen(),
+      home: const WelcomeScreen(),
     );
   }
 }
