+++ lib/user.model.dart
@@ -1,0 +1,8 @@
+class User {
+  late String userName;
+  late String phoneNumber;
+  late String email;
+  late String password;
+
+  User();
+}
